+++ src/scanpy/external/pp/_dca.py
@@ -11,7 +11,7 @@ if TYPE_CHECKING:
 
     from anndata import AnnData
 
-    from ..._compat import _LegacyRandom
+    from ..._utils.random import _LegacyRandom
 
     _AEType = Literal["zinb-conddisp", "zinb", "nb-conddisp", "nb"]
