+++ src/scanpy/external/pp/_magic.py
@@ -17,7 +17,7 @@ if TYPE_CHECKING:
 
     from anndata import AnnData
 
-    from ..._compat import _LegacyRandom
+    from ..._utils.random import _LegacyRandom
 
 MIN_VERSION = "2.0"
