+++ src/scanpy/external/tl/_phate.py
@@ -14,7 +14,7 @@ if TYPE_CHECKING:
 
     from anndata import AnnData
 
-    from ..._compat import _LegacyRandom
+    from ..._utils.random import _LegacyRandom
 
 
 @old_positionals(
