+++ src/scanpy/preprocessing/_deprecated/sampling.py
@@ -2,7 +2,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from ..._compat import _legacy_numpy_gen, old_positionals
+from ..._compat import old_positionals
+from ..._utils.random import legacy_numpy_gen
 from .._simple import sample
 
 if TYPE_CHECKING:
@@ -50,7 +51,7 @@ def subsample(
     returns a subsampled copy of it (`copy == True`).
 
     """
-    rng = _legacy_numpy_gen(random_state)
+    rng = legacy_numpy_gen(random_state)
     return sample(
         data=data, fraction=fraction, n=n_obs, rng=rng, copy=copy, replace=False, axis=0
     )
