+++ src/scanpy/preprocessing/_pca/__init__.py
@@ -27,8 +27,8 @@ if TYPE_CHECKING:
     import sklearn.decomposition as skld
     from numpy.typing import DTypeLike, NDArray
 
-    from ..._compat import _LegacyRandom
     from ..._utils import Empty
+    from ..._utils.random import _LegacyRandom
 
     MethodDaskML = type[dmld.PCA | dmld.IncrementalPCA | dmld.TruncatedSVD]
     MethodSklearn = type[skld.PCA | skld.TruncatedSVD]
