+++ src/scanpy/preprocessing/_recipes.py
@@ -16,7 +16,7 @@ from ._normalization import normalize_total
 if TYPE_CHECKING:
     from anndata import AnnData
 
-    from .._compat import _LegacyRandom
+    from .._utils.random import _LegacyRandom
 
 
 @old_positionals(
