+++ src/scanpy/tools/_score_genes.py
@@ -19,7 +19,7 @@ if TYPE_CHECKING:
     from anndata import AnnData
     from numpy.typing import DTypeLike, NDArray
 
-    from .._compat import _LegacyRandom
+    from .._utils.random import _LegacyRandom
 
     try:
         _StrIdx = pd.Index[str]
