+++ src/scanpy/tools/_umap.py
@@ -17,7 +17,7 @@ if TYPE_CHECKING:
 
     from anndata import AnnData
 
-    from .._compat import _LegacyRandom
+    from .._utils.random import _LegacyRandom
 
     _InitPos = Literal["paga", "spectral", "random"]
